rengeteg minden

+++ TaskFIFO.c
@@ -19,7 +19,7 @@
 typedef struct OneTimeTask {
     void (*func)(uint16);
     void (*funcnoparam)();
-    uint16 delay;
+    uint32 delay;
     uint16 data;  //A fuggveny parametere
     struct OneTimeTask *next;
 } OneTimeTask;
@@ -29,7 +29,7 @@ typedef struct TaskFIFO {
 } TaskFIFO;
 
 
-OneTimeTask * create_new_onetimetask( void (*func)(uint16), uint16 delay, uint16 data) {
+OneTimeTask * create_new_onetimetask( void (*func)(uint16), uint32 delay, uint16 data) {
     OneTimeTask * new_task;
     new_task = (OneTimeTask*)malloc(sizeof(OneTimeTask));
     if (new_task == NULL) {
@@ -46,7 +46,7 @@ OneTimeTask * create_new_onetimetask( void (*func)(uint16), uint16 delay, uint16
     return new_task;
 }
 
-OneTimeTask * create_new_noparam_onetimetask( void (*func)(), uint16 delay) {
+OneTimeTask * create_new_noparam_onetimetask( void (*func)(), uint32 delay) {
     OneTimeTask * new_task;
     new_task = (OneTimeTask*)malloc(sizeof(OneTimeTask));
     if (new_task == NULL) {
